Make float parsing work in different locales
Directly using atof is wrong as it is locale dependent and in some
languyages decimals separator is not . as is used in CCMX. This leads to
failure (or rather to trimming decimal part of the numbers).

This change uses printf to figure out what current locale use,
localeconv might sound as better to use, but it's definitely not thread
safe so it sounds like potential source of troubles.