Fix idendity device link with LINEARIZATION flags
This patch fixes a segmentation fault on Linux and likely other platforms.
For reproduction use:
$ linkicc -o device_link.icc -l `oyranos-profile --path sRGB.icc` `oyranos-profile --path sRGB.icc`

Discussion:
The crash happens if two identical profiles are linked to a
device link profile and cmsTransform2DeviceLink() obtains the
cmsFLAGS_CLUT_POST_LINEARIZATION or cmsFLAGS_CLUT_PRE_LINEARIZATION
flags as in the above demo. lcms2.h says to the flags:
define cmsFLAGS_CLUT_POST_LINEARIZATION  0x0001    // create postlinearization tables if possible
That means to my understanding, the flags shall perform a reasonable
fallback, in case the linearisation curves are not created. This patch
lets create a device link profile, even without the curves and
avoids a segmentation fault.