Reported bug from quangnh89
- Turns a warning into a fatal error.

I have tested the profile against a normal use and found despite this is
a real bug,  it is harmless in the security context. The out of bounds
condition can only be reached by building a special program.
Yet-existing, normal code are unaffected as it never needs to do this
particular sequence. I agree it is better to turn this warning into a
fatal error in order to keep safe.

NOT RELATED WITH SECURITY. IT CANNOT BE USED TO DO ANY EXPLOIT.