cmsalpha.c : fix swapped-endian case
Without this fix, FormatterPos() is incapable of handling swapped-endian requests (note that no case returns `2`, the index into the swapped-endian row!)

Fix is simple: when 2-bytes are requested, but half-float is not, check endianness.  If T_ENDIAN16, return 16SE; otherwise, return 16.

(Also a minor typo fix on a neighboring comment.)