Do not include autogenerated files in source control
Every time a user types 'make' in the source tree there are a huge number of
dirty files as the Fedora-generated ones are a tiny bit different from other
distros.

It also really messes up automated licence checks as the upstream is including
files with a very different licence to MIT.

Just include the correct magic in configure.ac so that all the files get
generated properly when running autogen.sh.

Also, actually compile the fast_float plugin rather than relying on a dummy
makefile to avoid an error.