Remove check on DLL when CMS_NO_REGISTER_KEYWORD is used
That was originally intended to prevent building DLLs with different ABI, but since you can actually build .so on linux with this flaw, makes no sense to check only one platform.

I would rather left the decision to developers. You are free to create a .so or DLL with incompatible parameter passing if you use CMS_NO_REGISTER_KEYWORD , it is up to you to check consistency.