Allow transform templates to be executed directly (it does nothing, though)
You may create transforms with formatters 0, that means the format is to be specified latter on. So, you were not supposed to call this template with cmsDoTransform before setting the format, but fuzzer did. So I am putting code to prevent this... without including a condition that  would ruin all cache and instruction branch guessing.