sanitize cgats parser
I don't know why people keeps fuzzing this code, but they do, and then they make a lot of noise. So let's make it less permissive and abort early when some wrong characters are found.
I apologize if someone got hurt in the process.
Otherwise this is harmless because is not used on ICC profile handling.