Fix PostScript output for linear tone curve
The Emit1Gamma function is expected to output a PostScript procedure, so
for a linear curve it will now generate an empty procedure to act as an
identity function.

Otherwise, subsequent PostScript output for a profile with linear TRC
may reference a procedure (for example, "lcms2gammaproc" for
EmitCIEBasedA) that is not actually defined.