revert creation of resources  that corrupted some PS interpreters
PS procedures, such as PaintProc, are not handled by the PS Interpreter until they are invoked. That means that the lcms2gammatable variable does not get defined in the current dictionary so it is unavailable for immediate variable replacement in lcms2gammaproc.  To fix that I reverted the whole lcms2gammatable thing.

The actual code generated is completely valid postscript. Those names were there to favor a specific interpreter, but unfortunately the price of that has revealed to be compatibility issues, so I cannot afford that.