Change computation of profile ID to follow actual ICC spec
From old spec 4.0.0 (ICC1-V40.pdf, ICC.1:2001-12) to actual one there is a breaking change that makes computation of profile ID different. This commit uses the latest spec 4.4.- The change is the requirement to set header flags to zero instead of attributes when computing the MD5 digest.
Thanks to Timo Kinnunen (@Overruler) for detecting and reporting the issue. This commit should fix #447