Fix adaptation issues when mixing V2 and V4 profiles in the pipeline
LCMS handles white point adaptation differently, depending on whether
the profile's white point is defined via cmsSigChromaticAdaptationTag
or cmsSigMediaWhitePointTag. If the white point is set via
cmsSigMediaWhitePointTag, LCMS uses "XYZ Scaling" method to adapt the
colors. If the white point is defined via cmsSigChromaticAdaptationTag,
then LCMS uses the embedded matrix (which is most likely "Bradford").

It causes some weird issues when combining profiles with white point
defined differently. For example, when color proofing a D65 device on
a D65 screen with absolute colorimetric intent, the D65->D50->D65
conversion may cause severe color degradations (instead of being identity).

The actual testcase looks like that:

* Proof profile: ICCv2, RGB "output" with D65 white point
  set via cmsSigMediaWhitePointTag

* Display profile: ICCv2, sRGB "display" with D65 white point set via
  cmsSigChromaticAdaptationTag (e.g. Elle's sRGB-v2 profile)

When combining these profiles in the pipeline, the conversions look
like that:

1) AToB1 is applied for the proof profile (D65 adaptation is
   converted into D50)
2) ComputeAbsoluteIntent adapts back D50 to D65 using "XYZ Scaling"
   method (because the output profile has only cmsSigMediaWhitePointTag)
3) ComputeAbsoluteIntent also applies also inverse adaptation of the
   display profile, which is "Bradfort", since the profile has
   cmsSigChromaticAdaptationTag.
4) BToA1 is applied for the display profile

In general, we convert a D65 space into D65 space with absolute intent,
so, theoretically, we expect steps 2 and 3 result in a noop. But, given
that the two profiles used a different adaptation method, the result
is not a noop and skews the colors into a bluish tint.

It looks like someone has already seen this issue with display class
profiles (more than 15 years ago) and added a special case for them,
which forces Bradfort matrix for cmsSigMediaWhitePointTag-only profiles.

This patch just extends this hack to the output-class profiles. Their
cmsSigMediaWhitePointTag is effectively converted into
cmsSigChromaticAdaptationTag with a Bradfort matrix.

PS:
This patch might a bigger discussion on how cmsSigChromaticAdaptationTag
should be handled. What happens if the user has two profiles with two
cmsSigChromaticAdaptationTag tags that use different adaptation methods?
Should CMS just discard the matrix and force the homogeneous method
for all the profiles?

PPS:
The patch also adds a unittest to verify the issue. Though I'm not
fully sure if 'output-v2-d65-no-chad.icc' is correct from ICC
specification point of view. I just used a modified version of
cmsCreateRGBProfile() to generate it. I wanted all the LUTs of
the profiles to be exactly the same to make the test possible.